+++ src/App.tsx
@@ -10,13 +10,13 @@ import { getCookie } from 'utils/ts/cookie';
 import AuthPage from 'pages/Auth/AuthPage';
 import LoginPage from 'pages/Auth/LoginPage';
 import BoardPage from 'pages/BoardPage';
-import StorePage from 'pages/BoardPage/Store/StorePage';
-import Notice from 'pages/BoardPage/Notice';
-import NoticePage from 'pages/BoardPage/Notice/NoticePage';
-import NoticeDetailPage from 'pages/BoardPage/Notice/NoticeDetailPage';
+import StorePage from 'pages/Store/StorePage';
+import NoticePage from 'pages/Notice/NoticePage';
+import NoticeListPage from 'pages/Notice/NoticeListPage';
+import NoticeDetailPage from 'pages/Notice/NoticeDetailPage';
 import Toast from 'components/common/Toast';
 import SignupPage from 'pages/Auth/SignupPage';
-import StoreDetailPage from 'pages/BoardPage/Store/StoreDetailPage';
+import StoreDetailPage from 'pages/Store/StoreDetailPage';
 import BusPage from 'pages/BusPage';
 
 const useTokenState = () => {
@@ -39,8 +39,8 @@ function App() {
           <Route path="/store" element={<StorePage />} />
           <Route path="/store/:id" element={<StoreDetailPage />} />
           <Route path="/bus" element={<BusPage />} />
-          <Route path="/board/notice" element={<Notice />}>
-            <Route path="/board/notice/" element={<NoticePage />} />
+          <Route path="/board/notice" element={<NoticePage />}>
+            <Route path="/board/notice/" element={<NoticeListPage />} />
             <Route path="/board/notice/:id" element={<NoticeDetailPage />} />
           </Route>
         </Route>
