+++ src/components/Post/PostDetailContent/index.tsx
@@ -6,13 +6,11 @@ type PostDetailContentProps = {
 
 function PostDetailContent(props: PostDetailContentProps) {
   const { content } = props;
-  // dangerouslySetInnerHTML에 대한 경고콘솔 제거
-  // eslint-disable-next-line
-  const contentParse = () => (<div dangerouslySetInnerHTML={{ __html: content }} />);
 
   return (
     <div className={styles.content}>
-      { contentParse() }
+      {/* eslint-disable-next-line  */}
+      <div dangerouslySetInnerHTML={{ __html: content }} /> 
     </div>
   );
 }
